[trello.com/c/1lX3h2xx] moved fetchBlockTimestamp to superclass
