+++ Adamant/Modules/Wallets/Ethereum/EthApiService.swift
@@ -58,3 +58,38 @@ class EthApiService: EthApiServiceProtocol, @unchecked Sendable {
         await api.service.setKeystoreManager(keystoreManager)
     }
 }
+
+extension EthApiService {
+    func fetchBlockTimestamp(blockNumberHex: String) async throws -> Date {
+        let body: [String: Any] = [
+            "jsonrpc": "2.0",
+            "method": "eth_getBlockByNumber",
+            "params": [blockNumberHex, false],
+            "id": 1
+        ]
+
+        let result: EthBlockResponse = try await requestApiCore(waitsForConnectivity: false) { core, origin in
+            await core.sendRequestJsonResponse(
+                origin: origin,
+                path: "",
+                method: .post,
+                jsonParameters: body
+            )
+        }.get()
+
+        guard
+            let timestampHex = result.result?.timestamp,
+            let timestampInt = UInt64(HexUtils.trimHexPrefix(from: timestampHex), radix: 16)
+        else {
+            throw WalletServiceError.remoteServiceError(message: "Invalid timestamp in block response")
+        }
+
+        return Date(timeIntervalSince1970: TimeInterval(timestampInt))
+    }
+}
+
+struct HexUtils {
+    static func trimHexPrefix(from hexString: String) -> String {
+        hexString.hasPrefix("0x") ? String(hexString.dropFirst(2)) : hexString
+    }
+}
